kospi200-futures-bot/requirements.txt
PyQt5>=5.15
loguru>=0.7
pandas>=2.0
numpy>=1.24
pyyaml>=6.0
